IntegratedProdList/requirements.txt
django==2.2
django-import-export
bcrypt
pandas
numpy